own package and included plugin parser
