+++ src/main/scala/de/halcony/appanalyzer/parser/AppAnalyzerParser.scala
@@ -1,0 +1,31 @@
+package de.halcony.appanalyzer.parser
+
+import de.halcony.appanalyzer.analysis.plugin.PluginManager
+import de.halcony.argparse.Parser
+
+class AppAnalyzerParser {
+  def createParser: Parser = {
+    createMainParser
+  }
+
+  private def createMainParser: Parser = {
+    Parser(
+      name = "AppAnalyzer",
+      description = "run apps and analyze their consent dialogs"
+    )
+      .addFlag(
+        name = "verbose",
+        short = "v",
+        description = "if set a stacktrace is provided with any fatal error"
+      )
+      .addOptional(
+        name = "config",
+        short = "c",
+        default = Some("config.json"),
+        description = "the configuration file"
+      )
+      .addSubparser(PluginManagerParser.createParser)
+      .addSubparser(DeletedAnalysisParser.createParser)
+      .addSubparser(RunParser.createParser)
+  }
+}
